Refactor scripts under trello subcommand
