+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/release/__init__.py
@@ -8,12 +8,11 @@ from .build import build
 from .changelog import changelog
 from .make import make
 from .show import show
-from .status import status
 from .tag import tag
-from .testable import testable
+from .trello import trello
 from .upload import upload
 
-ALL_COMMANDS = [build, changelog, make, show, status, tag, testable, upload]
+ALL_COMMANDS = [build, changelog, make, show, tag, trello, upload]
 
 
 @click.group(context_settings=CONTEXT_SETTINGS, short_help='Manage the release of checks')
